backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib==1.7.4
bcrypt==3.2.0
sqlalchemy
psycopg2-binary
alembic
python-multipart
requests
httpx

google-generativeai
jinja2
openai
anthropic
python-dotenv
Pillow
# For Image Generation
diffusers
transformers
accelerate
torch
websockets
aiofiles
websockets
pytest
pytest-asyncio